requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.4
celery==5.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
diff-match-patch==20241021
dj-database-url==3.1.2
Django==5.2
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-import-export==4.3.7
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.6.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.6.0
